requirements.txt
fastapi>=0.111.0
openai>=1.42.0
pydantic>=2.8.2
python-dotenv>=1.0.1
uvicorn>=0.30.0
pytest>=8.3.0
httpx>=0.28.1
